# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Chat Application API",
		"version": "1.0.0",
		"description": "API documentation for the real-time chat application"
	},
	"consumes": ["application/json", "multipart/form-data"],
	"produces": ["application/json", "multipart/form-data"],
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development server"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"tags": [
		{
			"name": "Auth",
			"description": "Authentication endpoints"
		},
		{
			"name": "Rooms",
			"description": "Chat room management"
		},
		{
			"name": "Messages",
			"description": "Message operations"
		}
	],
	"paths": {
		"/api/auth/register": {
			"post": {
				"summary": "Register a new user",
				"tags": ["Auth"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["username", "email", "password"],
								"properties": {
									"username": {
										"type": "string",
										"minLength": 3
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"minLength": 6
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User registered successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"email": {
													"type": "string"
												}
											}
										},
										"token": {
											"type": "string"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid input"
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"summary": "Login user",
				"tags": ["Auth"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["email", "password"],
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Login successful"
					},
					"401": {
						"description": "Invalid credentials"
					}
				}
			}
		},
		"/api/auth/me": {
			"get": {
				"summary": "Get current user profile",
				"tags": ["Auth"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User profile retrieved successfully"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
		"/api/rooms": {
			"get": {
				"summary": "Get all chat rooms",
				"tags": ["Rooms"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "List of chat rooms"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			},
			"post": {
				"summary": "Create a new chat room",
				"tags": ["Rooms"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["name", "topic"],
								"properties": {
									"name": {
										"type": "string"
									},
									"topic": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Chat room created"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
		"/api/messages/room/{roomId}": {
			"get": {
				"summary": "Get messages for a specific room",
				"tags": ["Messages"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "roomId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of messages"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
		"/api/messages/private": {
			"get": {
				"summary": "Get private messages for current user",
				"tags": ["Messages"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "List of private messages"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
		"/api/messages/unread": {
			"get": {
				"summary": "Get unread message count",
				"tags": ["Messages"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Number of unread messages"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
		"/api/messages/{messageId}/read": {
			"patch": {
				"summary": "Mark a message as read",
				"tags": ["Messages"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Message marked as read"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		}
	}
}